client: fix drive item updates inserting duplicates
The issue was that the detection in the drive component did not see the
contents of `paginationElem.items` because the ref was not properly
exposed. The value of the exposed property was a Proxy of an empty
array that was not a reference to the actual items array. Thus, when
checking if the item already exists or not, it never saw the item and
added it as if new.

closes https://akkoma.dev/FoundKeyGang/FoundKey/issues/367
closes https://akkoma.dev/FoundKeyGang/FoundKey/issues/368

Changelog: Fixed