fix MFM fg/bg color regex
CSS colors are either 3 or 6 hex digits, not 3 to 6 (which would allow
4 and 5 digit hex codes, which are not accepted).

Also adds an explicit null/undefined check.

Changes the default color for the $[bg ] function to something different
than the fg color so if you use both functions on a piece of text with
default values, the text stays somewhat readable.